refactor: Made logger private
