+++ cognite/powerops/resync/_logger.py
